==> 15_Create_Input_Rasters.R <==
#Create gridded input rasters using user-defined function 
#If you are looking for the file to makee
library(tiff)
library(raster)
library(rgdal)
library(lubridate)
library(MODIS)
library(ncdf4)
library(MODIStsp)

mean_na <- function(x) {
  mean(x,na.rm=T)
}

Global_LC <- raster("F:/Upscaling_Project/Gridded_Inputs/LC_5min_global_2001.tif")
Upscext <- extent(EVI)
EVI <- (raster("F:/Upscaling_Project/Gridded_Inputs/Monthly_EVI/Apr_2001_EVI.tif"))

LC_Resample <- resample(Global_LC, EVI, method="bilinear")
plot(LC_Resample)

# Elevation
dem = raster("F:/Upscaling_Project/Gridded_Inputs/Hydro1k/na_dem.bil")
dem[dem==-9999] = NA
dem[dem>10000] = NA
plot(dem)
EVI <- (raster("F:/Upscaling_Project/Gridded_Inputs/Monthly_EVI/Apr_2001_EVI.tif"))
Upscext <- extent(EVI)
temp = projectRaster(dem, EVI) #reproject/resample to MODIS extent and resolution
temp <- writeRaster(temp, filename="F:/Upscaling_Project/Gridded_Inputs/Hydro1k/SW_dem.tif",format="GTiff",overwrite=TRUE)


writeRaster(LC_Resample, "F:/Upscaling_Project/Gridded_Inputs/LC_Resample.tif")

swe <-stack("D:/Upscaling_Project/Gridded_Inputs/Daymet/upscalingArea_DAYMET_swe_2000_2016_AOI.tif")
srad <-stack("D:/Upscaling_Project/Gridded_Inputs/Daymet/upscalingArea_DAYMET_srad_2000_2016_AOI.tif")
plot(srad)
plot(swe)

create_input_raster <- function(band1, month, monthno, year){
  filenameEVI <- paste0("F:/Upscaling_Project/Gridded_Inputs/Monthly_EVI/", month, "_", year, "_", "EVI.tif")
  EVI <- raster(filenameEVI)
  filenamePrecip <- "F:/Upscaling_Project/Gridded_Inputs/Daymet/upscalingArea_DAYMET_prcp_2000_2016_AOI.tif"
  filenameTmax <- "F:/Upscaling_Project/Gridded_Inputs/Daymet/upscalingArea_DAYMET_tmax_2000_2016_AOI.tif"
  filenameTmin <- "F:/Upscaling_Project/Gridded_Inputs/Daymet/upscalingArea_DAYMET_tmin_2000_2016_AOI.tif"
  DEM <- "F:/Upscaling_Project/Gridded_Inputs/Hydro1k/SW_dem.tif"
  
    Precip <- raster(filenamePrecip, band = band1)
  Tmax <- raster(filenameTmax, band = band1)
  Tmin <- raster(filenameTmin, band = band1)
  
  Tmax[Tmax==-9999] <-NA
  Tmin[Tmin==-9999] <-NA
  Precip[Precip==-9999] <-NA
  
  Upscext <- extent(EVI)
  
  Precipresample <- resample(Precip, EVI, method="bilinear")
  Tmaxresample <- resample(Tmax, EVI, method="bilinear")
  Tminresample <- resample(Tmin, EVI, method="bilinear")
  
  monthrast = raster (ext=Upscext, res=0.002081004)
  values(monthrast) <- monthno
  
  rast_stack <- stack(Tminresample, Tmaxresample, Precipresample, EVI, monthrast)
  names(rast_stack) <- paste(c("tmin", "tmax", "precip", "NDVI", "month", "DEM"))
  
  outputfilename <- paste("F:/Upscaling_Project/Gridded_inputs/RF_Input/",month,"_",year,".tif")
  writeRaster(rast_stack, outputfilename)
  return(rast_stack)
}


dem <- raster("F:/Upscaling_Project/Gridded_Inputs/Hydro1k/SW_dem.tif")

#For 2001-----------------------------------
create_input_raster(band1=25, month="Jan", monthno=1, year=2001)
create_input_raster(band1=26, month="Feb", monthno=2, year=2001)
create_input_raster(band1=27, month="Mar", monthno=3, year=2001)
create_input_raster(band1=28, month="Apr", monthno=4, year=2001)
create_input_raster(band1=29, month="May", monthno=5, year=2001)
create_input_raster(band1=30, month="Jun", monthno=6, year=2001)
create_input_raster(band1=31, month="Jul", monthno=7, year=2001)
create_input_raster(band1=32, month="Aug", monthno=8, year=2001)
create_input_raster(band1=33, month="Sep", monthno=9, year=2001)
create_input_raster(band1=34, month="Oct", monthno=10, year=2001)
create_input_raster(band1=35, month="Nov", monthno=11, year=2001)
create_input_raster(band1=36, month="Dec", monthno=12, year=2001)

Jan_2001 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2001.tif", dem)
plot(Jan_2001)
names(Jan_2001) <- paste(c("tmin", "tmax", "precip", "NDVI", "month", "elev"))
Feb_2002 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2002.tif")
names(Feb_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2002.tif")
names(Mar_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2002.tif")
names(Apr_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2002.tif")
names(May_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2002.tif")
names(Jun_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2002.tif")
names(Jul_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2002.tif")
names(Aug_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2002.tif")
names(Sep_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2002.tif")
names(Oct_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2002<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2002.tif")
names(Nov_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2002.tif")
names(Dec_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
#For 2002-----------------------------------
create_input_raster(band1=25, month="Jan", monthno=1, year=2002)
create_input_raster(band1=26, month="Feb", monthno=2, year=2002)
create_input_raster(band1=27, month="Mar", monthno=3, year=2002)
create_input_raster(band1=28, month="Apr", monthno=4, year=2002)
create_input_raster(band1=29, month="May", monthno=5, year=2002)
create_input_raster(band1=30, month="Jun", monthno=6, year=2002)
create_input_raster(band1=31, month="Jul", monthno=7, year=2002)
create_input_raster(band1=32, month="Aug", monthno=8, year=2002)
create_input_raster(band1=33, month="Sep", monthno=9, year=2002)
create_input_raster(band1=34, month="Oct", monthno=10, year=2002)
create_input_raster(band1=35, month="Nov", monthno=11, year=2002)
create_input_raster(band1=36, month="Dec", monthno=12, year=2002)

Jan_2002 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2002.tif")
names(Jan_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2002 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2002.tif")
names(Feb_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2002.tif")
names(Mar_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2002.tif")
names(Apr_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2002.tif")
names(May_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2002.tif")
names(Jun_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2002.tif")
names(Jul_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2002.tif")
names(Aug_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2002.tif")
names(Sep_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2002.tif")
names(Oct_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2002<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2002.tif")
names(Nov_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2002<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2002.tif")
names(Dec_2002) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2003---------
create_input_raster(band1=37, month="Jan", monthno=1, year=2003)
create_input_raster(band1=38, month="Feb", monthno=2, year=2003)
create_input_raster(band1=39, month="Mar", monthno=3, year=2003)
create_input_raster(band1=40, month="Apr", monthno=4, year=2003)
create_input_raster(band1=41, month="May", monthno=5, year=2003)
create_input_raster(band1=42, month="Jun", monthno=6, year=2003)
create_input_raster(band1=43, month="Jul", monthno=7, year=2003)
create_input_raster(band1=44, month="Aug", monthno=8, year=2003)
create_input_raster(band1=45, month="Sep", monthno=9, year=2003)
create_input_raster(band1=46, month="Oct", monthno=10, year=2003)
create_input_raster(band1=47, month="Nov", monthno=11, year=2003)
create_input_raster(band1=48, month="Dec", monthno=12, year=2003)

Jan_2003 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2003.tif")
names(Jan_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2003 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2003.tif")
names(Feb_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2003<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2003.tif")
names(Mar_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2003<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2003.tif")
names(Apr_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2003<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2003.tif")
names(May_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2003<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2003.tif")
names(Jun_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2003<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2003.tif")
names(Jul_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2003<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2003.tif")
names(Aug_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2003<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2003.tif")
names(Sep_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2003<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2003.tif")
names(Oct_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2003<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2003.tif")
names(Nov_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2003<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2003.tif")
names(Dec_2003) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2004--------
create_input_raster(band1=49, month="Jan", monthno=1, year=2004)
create_input_raster(band1=50, month="Feb", monthno=2, year=2004)
create_input_raster(band1=51, month="Mar", monthno=3, year=2004)
create_input_raster(band1=52, month="Apr", monthno=4, year=2004)
create_input_raster(band1=53, month="May", monthno=5, year=2004)
create_input_raster(band1=54, month="Jun", monthno=6, year=2004)
create_input_raster(band1=55, month="Jul", monthno=7, year=2004)
create_input_raster(band1=56, month="Aug", monthno=8, year=2004)
create_input_raster(band1=57, month="Sep", monthno=9, year=2004)
create_input_raster(band1=58, month="Oct", monthno=10, year=2004)
create_input_raster(band1=59, month="Nov", monthno=11, year=2004)
create_input_raster(band1=60, month="Dec", monthno=12, year=2004)

Jan_2004 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2004.tif")
names(Jan_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2004 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2004.tif")
names(Feb_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2004<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2004.tif")
names(Mar_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2004<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2004.tif")
names(Apr_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2004<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2004.tif")
names(May_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2004<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2004.tif")
names(Jun_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2004<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2004.tif")
names(Jul_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2004<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2004.tif")
names(Aug_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2004<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2004.tif")
names(Sep_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2004<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2004.tif")
names(Oct_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2004<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2004.tif")
names(Nov_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2004<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2004.tif")
names(Dec_2004) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2005--------
create_input_raster(band1=61, month="Jan", monthno=1, year=2005)
create_input_raster(band1=62, month="Feb", monthno=2, year=2005)
create_input_raster(band1=63, month="Mar", monthno=3, year=2005)
create_input_raster(band1=64, month="Apr", monthno=4, year=2005)
create_input_raster(band1=65, month="May", monthno=5, year=2005)
create_input_raster(band1=66, month="Jun", monthno=6, year=2005)
create_input_raster(band1=67, month="Jul", monthno=7, year=2005)
create_input_raster(band1=68, month="Aug", monthno=8, year=2005)
create_input_raster(band1=69, month="Sep", monthno=9, year=2005)
create_input_raster(band1=70, month="Oct", monthno=10, year=2005)
create_input_raster(band1=71, month="Nov", monthno=11, year=2005)
create_input_raster(band1=72, month="Dec", monthno=12, year=2005)

Jan_2005 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2005.tif")
names(Jan_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2005 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2005.tif")
names(Feb_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2005<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2005.tif")
names(Mar_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2005<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2005.tif")
names(Apr_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2005<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2005.tif")
names(May_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2005<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2005.tif")
names(Jun_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2005<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2005.tif")
names(Jul_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2005<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2005.tif")
names(Aug_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2005<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2005.tif")
names(Sep_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2005<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2005.tif")
names(Oct_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2005<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2005.tif")
names(Nov_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2005<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2005.tif")
names(Dec_2005) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2006---------
create_input_raster(band1=73, month="Jan", monthno=1, year=2006)
create_input_raster(band1=74, month="Feb", monthno=2, year=2006)
create_input_raster(band1=75, month="Mar", monthno=3, year=2006)
create_input_raster(band1=76, month="Apr", monthno=4, year=2006)
create_input_raster(band1=77, month="May", monthno=5, year=2006)
create_input_raster(band1=78, month="Jun", monthno=6, year=2006)
create_input_raster(band1=79, month="Jul", monthno=7, year=2006)
create_input_raster(band1=80, month="Aug", monthno=8, year=2006)
create_input_raster(band1=81, month="Sep", monthno=9, year=2006)
create_input_raster(band1=82, month="Oct", monthno=10, year=2006)
create_input_raster(band1=83, month="Nov", monthno=11, year=2006)
create_input_raster(band1=84, month="Dec", monthno=12, year=2006)

Jan_2006 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2006.tif")
names(Jan_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2006 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2006.tif")
names(Feb_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2006<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2006.tif")
names(Mar_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2006<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2006.tif")
names(Apr_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2006<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2006.tif")
names(May_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2006<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2006.tif")
names(Jun_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2006<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2006.tif")
names(Jul_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2006<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2006.tif")
names(Aug_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2006<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2006.tif")
names(Sep_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2006<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2006.tif")
names(Oct_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2006<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2006.tif")
names(Nov_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2006<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2006.tif")
names(Dec_2006) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2007---------
create_input_raster(band1=85, month="Jan", monthno=1, year=2007)
create_input_raster(band1=86, month="Feb", monthno=2, year=2007)
create_input_raster(band1=87, month="Mar", monthno=3, year=2007)
create_input_raster(band1=88, month="Apr", monthno=4, year=2007)
create_input_raster(band1=89, month="May", monthno=5, year=2007)
create_input_raster(band1=90, month="Jun", monthno=6, year=2007)
create_input_raster(band1=91, month="Jul", monthno=7, year=2007)
create_input_raster(band1=92, month="Aug", monthno=8, year=2007)
create_input_raster(band1=93, month="Sep", monthno=9, year=2007)
create_input_raster(band1=94, month="Oct", monthno=10, year=2007)
create_input_raster(band1=95, month="Nov", monthno=11, year=2007)
create_input_raster(band1=96, month="Dec", monthno=12, year=2007)

Jan_2007 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2007.tif")
names(Jan_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2007 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2007.tif")
names(Feb_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2007<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2007.tif")
names(Mar_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2007<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2007.tif")
names(Apr_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2007<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2007.tif")
names(May_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2007<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2007.tif")
names(Jun_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2007<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2007.tif")
names(Jul_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2007<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2007.tif")
names(Aug_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2007<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2007.tif")
names(Sep_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2007<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2007.tif")
names(Oct_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2007<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2007.tif")
names(Nov_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2007<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2007.tif")
names(Dec_2007) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2008---------
create_input_raster(band1=97, month="Jan", monthno=1, year=2008)
create_input_raster(band1=98, month="Feb", monthno=2, year=2008)
create_input_raster(band1=99, month="Mar", monthno=3, year=2008)
create_input_raster(band1=100, month="Apr", monthno=4, year=2008)
create_input_raster(band1=101, month="May", monthno=5, year=2008)
create_input_raster(band1=102, month="Jun", monthno=6, year=2008)
create_input_raster(band1=103, month="Jul", monthno=7, year=2008)
create_input_raster(band1=104, month="Aug", monthno=8, year=2008)
create_input_raster(band1=105, month="Sep", monthno=9, year=2008)
create_input_raster(band1=106, month="Oct", monthno=10, year=2008)
create_input_raster(band1=107, month="Nov", monthno=11, year=2008)
create_input_raster(band1=108, month="Dec", monthno=12, year=2008)


Jan_2008 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2008.tif")
names(Jan_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2008 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2008.tif")
names(Feb_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2008<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2008.tif")
names(Mar_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2008<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2008.tif")
names(Apr_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2008<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2008.tif")
names(May_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2008<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2008.tif")
names(Jun_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2008<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2008.tif")
names(Jul_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2008<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2008.tif")
names(Aug_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2008<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2008.tif")
names(Sep_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2008<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2008.tif")
names(Oct_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2008<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2008.tif")
names(Nov_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2008<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2008.tif")
names(Dec_2008) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))


#For 2009---------
create_input_raster(band1=109, month="Jan", monthno=1, year=2009)
create_input_raster(band1=110, month="Feb", monthno=2, year=2009)
create_input_raster(band1=111, month="Mar", monthno=3, year=2009)
create_input_raster(band1=112, month="Apr", monthno=4, year=2009)
create_input_raster(band1=113, month="May", monthno=5, year=2009)
create_input_raster(band1=114, month="Jun", monthno=6, year=2009)
create_input_raster(band1=115, month="Jul", monthno=7, year=2009)
create_input_raster(band1=116, month="Aug", monthno=8, year=2009)
create_input_raster(band1=117, month="Sep", monthno=9, year=2009)
create_input_raster(band1=118, month="Oct", monthno=10, year=2009)
create_input_raster(band1=119, month="Nov", monthno=11, year=2009)
create_input_raster(band1=120, month="Dec", monthno=12, year=2009)

Jan_2009 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2009.tif")
names(Jan_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2009 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2009.tif")
names(Feb_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2009<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2009.tif")
names(Mar_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2009<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2009.tif")
names(Apr_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2009<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2009.tif")
names(May_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2009<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2009.tif")
names(Jun_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2009<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2009.tif")
names(Jul_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2009<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2009.tif")
names(Aug_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2009<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2009.tif")
names(Sep_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2009<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2009.tif")
names(Oct_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2009<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2009.tif")
names(Nov_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2009<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2009.tif")
names(Dec_2009) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2010---------
create_input_raster(band1=121, month="Jan", monthno=1, year=2010)
create_input_raster(band1=122, month="Feb", monthno=2, year=2010)
create_input_raster(band1=123, month="Mar", monthno=3, year=2010)
create_input_raster(band1=124, month="Apr", monthno=4, year=2010)
create_input_raster(band1=125, month="May", monthno=5, year=2010)
create_input_raster(band1=126, month="Jun", monthno=6, year=2010)
create_input_raster(band1=127, month="Jul", monthno=7, year=2010)
create_input_raster(band1=128, month="Aug", monthno=8, year=2010)
create_input_raster(band1=129, month="Sep", monthno=9, year=2010)
create_input_raster(band1=130, month="Oct", monthno=10, year=2010)
create_input_raster(band1=131, month="Nov", monthno=11, year=2010)
create_input_raster(band1=132, month="Dec", monthno=12, year=2010)

Jan_2010 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2010.tif")
names(Jan_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2010 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2010.tif")
names(Feb_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2010<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2010.tif")
names(Mar_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2010<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2010.tif")
names(Apr_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2010<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2010.tif")
names(May_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2010<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2010.tif")
names(Jun_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2010<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2010.tif")
names(Jul_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2010<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2010.tif")
names(Aug_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2010<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2010.tif")
names(Sep_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2010<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2010.tif")
names(Oct_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2010<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2010.tif")
names(Nov_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2010<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2010.tif")
names(Dec_2010) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2011--------
create_input_raster(band1=133, month="Jan", monthno=1, year=2011)
create_input_raster(band1=134, month="Feb", monthno=2, year=2011)
create_input_raster(band1=135, month="Mar", monthno=3, year=2011)
create_input_raster(band1=136, month="Apr", monthno=4, year=2011)
create_input_raster(band1=137, month="May", monthno=5, year=2011)
create_input_raster(band1=138, month="Jun", monthno=6, year=2011)
create_input_raster(band1=139, month="Jul", monthno=7, year=2011)
create_input_raster(band1=140, month="Aug", monthno=8, year=2011)
create_input_raster(band1=141, month="Sep", monthno=9, year=2011)
create_input_raster(band1=142, month="Oct", monthno=10, year=2011)
create_input_raster(band1=143, month="Nov", monthno=11, year=2011)
create_input_raster(band1=144, month="Dec", monthno=12, year=2011)

Jan_2011 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2011.tif")
names(Jan_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2011 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2011.tif")
names(Feb_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2011<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2011.tif")
names(Mar_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2011<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2011.tif")
names(Apr_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2011<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2011.tif")
names(May_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2011<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2011.tif")
names(Jun_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2011<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2011.tif")
names(Jul_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2011<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2011.tif")
names(Aug_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2011<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2011.tif")
names(Sep_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2011<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2011.tif")
names(Oct_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2011<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2011.tif")
names(Nov_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2011<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2011.tif")
names(Dec_2011) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2012---------
create_input_raster(band1=145, month="Jan", monthno=1, year=2012)
create_input_raster(band1=146, month="Feb", monthno=2, year=2012)
create_input_raster(band1=147, month="Mar", monthno=3, year=2012)
create_input_raster(band1=148, month="Apr", monthno=4, year=2012)
create_input_raster(band1=149, month="May", monthno=5, year=2012)
create_input_raster(band1=150, month="Jun", monthno=6, year=2012)
create_input_raster(band1=151, month="Jul", monthno=7, year=2012)
create_input_raster(band1=152, month="Aug", monthno=8, year=2012)
create_input_raster(band1=153, month="Sep", monthno=9, year=2012)
create_input_raster(band1=154, month="Oct", monthno=10, year=2012)
create_input_raster(band1=155, month="Nov", monthno=11, year=2012)
create_input_raster(band1=156, month="Dec", monthno=12, year=2012)


Jan_2012 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2012.tif")
names(Jan_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2012 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2012.tif")
names(Feb_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2012<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2012.tif")
names(Mar_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2012<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2012.tif")
names(Apr_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2012<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2012.tif")
names(May_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2012<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2012.tif")
names(Jun_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2012<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2012.tif")
names(Jul_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2012<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2012.tif")
names(Aug_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2012<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2012.tif")
names(Sep_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2012<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2012.tif")
names(Oct_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2012<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2012.tif")
names(Nov_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2012<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2012.tif")
names(Dec_2012) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2013---------

create_input_raster(band1=157, month="Jan", monthno=1, year=2013)
create_input_raster(band1=158, month="Feb", monthno=2, year=2013)
create_input_raster(band1=159, month="Mar", monthno=3, year=2013)
create_input_raster(band1=160, month="Apr", monthno=4, year=2013)
create_input_raster(band1=161, month="May", monthno=5, year=2013)
create_input_raster(band1=162, month="Jun", monthno=6, year=2013)
create_input_raster(band1=163, month="Jul", monthno=7, year=2013)
create_input_raster(band1=164, month="Aug", monthno=8, year=2013)
create_input_raster(band1=165, month="Sep", monthno=9, year=2013)
create_input_raster(band1=166, month="Oct", monthno=10, year=2013)
create_input_raster(band1=167, month="Nov", monthno=11, year=2013)
create_input_raster(band1=168, month="Dec", monthno=12, year=2013)

Jan_2013 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2013.tif")
names(Jan_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2013 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2013.tif")
names(Feb_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2013<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2013.tif")
names(Mar_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2013<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2013.tif")
names(Apr_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2013<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2013.tif")
names(May_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2013<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2013.tif")
names(Jun_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2013<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2013.tif")
names(Jul_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2013<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2013.tif")
names(Aug_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2013<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2013.tif")
names(Sep_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2013<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2013.tif")
names(Oct_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2013<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2013.tif")
names(Nov_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2013<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2013.tif")
names(Dec_2013) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2014--------


create_input_raster(band1=169, month="Jan", monthno=1, year=2014)
create_input_raster(band1=170, month="Feb", monthno=2, year=2014)
create_input_raster(band1=171, month="Mar", monthno=3, year=2014)
create_input_raster(band1=172, month="Apr", monthno=4, year=2014)
create_input_raster(band1=173, month="May", monthno=5, year=2014)
create_input_raster(band1=174, month="Jun", monthno=6, year=2014)
create_input_raster(band1=175, month="Jul", monthno=7, year=2014)
create_input_raster(band1=176, month="Aug", monthno=8, year=2014)
create_input_raster(band1=177, month="Sep", monthno=9, year=2014)
create_input_raster(band1=178, month="Oct", monthno=10, year=2014)
create_input_raster(band1=179, month="Nov", monthno=11, year=2014)
create_input_raster(band1=180, month="Dec", monthno=12, year=2014)

Jan_2014 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2014.tif")
names(Jan_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2014 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2014.tif")
names(Feb_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2014<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2014.tif")
names(Mar_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2014<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2014.tif")
names(Apr_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2014<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2014.tif")
names(May_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2014<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2014.tif")
names(Jun_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2014<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2014.tif")
names(Jul_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2014<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2014.tif")
names(Aug_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2014<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2014.tif")
names(Sep_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2014<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2014.tif")
names(Oct_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2014<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2014.tif")
names(Nov_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2014<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2014.tif")
names(Dec_2014) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2015---------
create_input_raster(band1=181, month="Jan", monthno=1, year=2015)
create_input_raster(band1=182, month="Feb", monthno=2, year=2015)
create_input_raster(band1=183, month="Mar", monthno=3, year=2015)
create_input_raster(band1=184, month="Apr", monthno=4, year=2015)
create_input_raster(band1=185, month="May", monthno=5, year=2015)
create_input_raster(band1=186, month="Jun", monthno=6, year=2015)
create_input_raster(band1=187, month="Jul", monthno=7, year=2015)
create_input_raster(band1=188, month="Aug", monthno=8, year=2015)
create_input_raster(band1=189, month="Sep", monthno=9, year=2015)
create_input_raster(band1=190, month="Oct", monthno=10, year=2015)
create_input_raster(band1=191, month="Nov", monthno=11, year=2015)
create_input_raster(band1=192, month="Dec", monthno=12, year=2015)

Jan_2015 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2015.tif")
names(Jan_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2015 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2015.tif")
names(Feb_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2015<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2015.tif")
names(Mar_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2015<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2015.tif")
names(Apr_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2015<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2015.tif")
names(May_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2015<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2015.tif")
names(Jun_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2015<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2015.tif")
names(Jul_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2015<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2015.tif")
names(Aug_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2015<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2015.tif")
names(Sep_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2015<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2015.tif")
names(Oct_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2015<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2015.tif")
names(Nov_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2015<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2015.tif")
names(Dec_2015) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))

#For 2016--------
create_input_raster(band1=193, month="Jan", monthno=1, year=2016)
create_input_raster(band1=194, month="Feb", monthno=2, year=2016)
create_input_raster(band1=195, month="Mar", monthno=3, year=2016)
create_input_raster(band1=196, month="Apr", monthno=4, year=2016)
create_input_raster(band1=197, month="May", monthno=5, year=2016)
create_input_raster(band1=198, month="Jun", monthno=6, year=2016)
create_input_raster(band1=199, month="Jul", monthno=7, year=2016)
create_input_raster(band1=200, month="Aug", monthno=8, year=2016)
create_input_raster(band1=201, month="Sep", monthno=9, year=2016)
create_input_raster(band1=202, month="Oct", monthno=10, year=2016)
create_input_raster(band1=203, month="Nov", monthno=11, year=2016)
create_input_raster(band1=204, month="Dec", monthno=12, year=2016)

Jan_2016 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jan_2016.tif")
names(Jan_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Feb_2016 <- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Feb_2016.tif")
names(Feb_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Mar_2016<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Mar_2016.tif")
names(Mar_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Apr_2016<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Apr_2016.tif")
names(Apr_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
May_2016<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/May_2016.tif")
names(May_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jun_2016<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jun_2016.tif")
names(Jun_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Jul_2016<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Jul_2016.tif")
names(Jul_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Aug_2016<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Aug_2016.tif")
names(Aug_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Sep_2016<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Sep_2016.tif")
names(Sep_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Oct_2016<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Oct_2016.tif")
names(Oct_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Nov_2016<-stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Nov_2016.tif")
names(Nov_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
Dec_2016<- stack("F:/Upscaling_Project/Gridded_Inputs/RF_Input/Dec_2016.tif")
names(Dec_2016) <- paste(c("tmin", "tmax", "precip", "NDVI", "month"))
